deepgram-transcription-agent/requirements.txt
livekit
livekit-agents
python-dotenv
livekit-plugins-deepgram